API/requirements.txt
async-timeout==4.0.2
build==0.10.0
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.14.0
django-extensions==3.2.1
django-jsonfield==1.4.1
djangorestframework==3.14.0
dnspython==1.16.0
eventlet==0.21.0
Pillow==9.5.0
protobuf==3.20.3
psycopg2-binary==2.9.6
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
#python-engineio==4.4.0
python-socketio==5.8.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
SQLAlchemy==1.4.47
sqlparse==0.4.4
